Move pkg/subctl/operator/servicediscoverycr under pkg
