+++ pkg/deploy/servicediscovery.go
@@ -24,7 +24,7 @@ import (
 	"github.com/submariner-io/admiral/pkg/reporter"
 	"github.com/submariner-io/subctl/internal/constants"
 	"github.com/submariner-io/subctl/pkg/broker"
-	"github.com/submariner-io/subctl/pkg/subctl/operator/servicediscoverycr"
+	"github.com/submariner-io/subctl/pkg/servicediscoverycr"
 	submariner "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/pkg/client"
 	v1 "k8s.io/api/core/v1"
