+++ pkg/servicediscoverycr/ensure.go
